RFC: Add COUNT option with loop to onSpinWait
The onSpinWait implementation on Aarch64 currently provides a means to
insert a variable number of ISB, YIELD or NOP instructions into the
onSpinWait and SpinPause implementations. However, the YIELD instruction
is really only applicable for possible SMT enabled cores, NOPs tend to
have little effect, and ISBs are not intended for delays, and their
delay can be expected in future.

This patch provides an additional "counter" option that uses the
CNTVCT_EL0 register, which increments at the rate set in the CNTFRQ_EL0
register. Using this, a delay can be inserted that is reasonably
predictable for most architectures.

The options are:
  -XX:OnSpinWaitInst=counter
  -XX:OnSpinWaitCounterDelay=<integer>

Change-Id: I1dd11739a779539d738e36d8064d343815f4db96
CustomizedGitHooks: yes